requirements.txt
absl-py==1.0.0
numpy==1.21.4
tflite-runtime==2.7.0
opencv-python==4.5.4.60